requirements.txt
astroid==2.2.5
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-filter==2.1.0
django-heroku==0.3.1
django-phonenumber-field==2.3.1
djangorestframework==3.9.2
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
phonenumbers==8.10.10
pkg-resources==0.0.0
psycopg2==2.8.2
pylint==2.3.1
python-decouple==3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
urllib3==1.24.2
waitress==1.3.0
whitenoise==4.1.2
wrapt==1.11.1
